+++ src/test/java/kr/co/pawong/pwbe/infrastructure/ai/huggingface/HuggingFaceChatAdapterTest.java
@@ -1,6 +1,6 @@
-package kr.co.pawong.pwbe.adoption.infrastructure.external.ai.openai;
+package kr.co.pawong.pwbe.infrastructure.ai.huggingface;
 
-import kr.co.pawong.pwbe.infrastructure.ai.openai.OpenAiChatAdapter;
+import kr.co.pawong.pwbe.infrastructure.ai.huggingface.HuggingFaceChatAdapter;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,10 +10,10 @@ import java.util.List;
 
 @Disabled("AI 연동 테스트는 전체 빌드 시 제외")
 @SpringBootTest(properties = "spring.profiles.active=dev")
-class OpenAiChatAdapterTest {
+class HuggingFaceChatAdapterTest {
 
     @Autowired
-    private OpenAiChatAdapter chatAdapter;
+    private HuggingFaceChatAdapter chatAdapter;
 
     @Test
     void LLM_질의() {
@@ -30,9 +30,8 @@ class OpenAiChatAdapterTest {
 
     @Test
     void 태깅() {
-        String animalFeature = "새낑 강아지인데 어르신들과 잘 지냄. 구조? 순종적인? / 체중: 7kg / 색: 검정 / 나이 9살";
-        List<String> output = chatAdapter.getTagsByFeature(animalFeature);
+        String searchTerm = "새낑 강아지인데 어르신들과 잘 지냄. 구조? 순종적인?";
+        List<String> output = chatAdapter.getTagsByFeature(searchTerm);
         System.out.println(output);
-        // 예상 출력: [노인과잘지냄, 무게가중간임, 색깔이어두움, 나이가많음, 구조됨, 온순함]
     }
 }
