+++ src/test/java/kr/co/pawong/pwbe/infrastructure/ai/openai/OpenAiEmbeddingAdapterTest.java
@@ -1,4 +1,4 @@
-package kr.co.pawong.pwbe.adoption.infrastructure.external.ai.openai;
+package kr.co.pawong.pwbe.infrastructure.ai.openai;
 
 import kr.co.pawong.pwbe.infrastructure.ai.openai.OpenAiEmbeddingAdapter;
 import org.junit.jupiter.api.Disabled;
@@ -11,7 +11,7 @@ import java.util.Arrays;
 /**
  * infrastructure라.. 스프링 부트 없이 테스트가 안 될 것 같아요. 비활성화 해놓겠습니다.
  */
-//@Disabled("AI 연동 테스트는 전체 빌드 시 제외")
+@Disabled("AI 연동 테스트는 전체 빌드 시 제외")
 @SpringBootTest(properties = "spring.profiles.active=dev")
 class OpenAiEmbeddingAdapterTest {
